Prevent warning on possible value-change

+++ include/nonstd/variant.hpp
@@ -649,10 +649,10 @@ union max_align_t
 #define variant_ALIGN_AS( to_align ) \
     typename detail::type_of_size< detail::alignment_types, detail::alignment_of< to_align >::value >::type
 
-template <typename T>
+template< typename T >
 struct alignment_of;
 
-template <typename T>
+template< typename T >
 struct alignment_of_hack
 {
     char c;
@@ -660,7 +660,7 @@ struct alignment_of_hack
     alignment_of_hack();
 };
 
-template <unsigned A, unsigned S>
+template< size_t A, size_t S >
 struct alignment_logic
 {
     enum { value = A < S ? A : S };
@@ -792,7 +792,7 @@ struct helper
         case 13: as<T13>( data )->~T13(); break;
         case 14: as<T14>( data )->~T14(); break;
         case 15: as<T15>( data )->~T15(); break;
-        
+
         }
     }
 
@@ -835,7 +835,7 @@ struct helper
         case 13: new( to_value ) T13( std::forward<T13>( *as<T13>( from_value ) ) ); break;
         case 14: new( to_value ) T14( std::forward<T14>( *as<T14>( from_value ) ) ); break;
         case 15: new( to_value ) T15( std::forward<T15>( *as<T15>( from_value ) ) ); break;
-        
+
         }
         return to_index_t( from_index );
     }
@@ -861,7 +861,7 @@ struct helper
         case 13: new( to_value ) T13( *as<T13>( from_value ) ); break;
         case 14: new( to_value ) T14( *as<T14>( from_value ) ); break;
         case 15: new( to_value ) T15( *as<T15>( from_value ) ); break;
-        
+
         }
         return to_index_t( from_index );
     }
@@ -991,7 +991,7 @@ public:
     variant( T13 const & t13 ) : type_index( 13 ) { new( ptr() ) T13( t13 ); }
     variant( T14 const & t14 ) : type_index( 14 ) { new( ptr() ) T14( t14 ); }
     variant( T15 const & t15 ) : type_index( 15 ) { new( ptr() ) T15( t15 ); }
-    
+
 
 #if variant_CPP11_OR_GREATER
     variant( T0 && t0 ) : type_index( 0 ) { new( ptr() ) T0( std::move(t0) ); }
@@ -1010,7 +1010,7 @@ public:
     variant( T13 && t13 ) : type_index( 13 ) { new( ptr() ) T13( std::move(t13) ); }
     variant( T14 && t14 ) : type_index( 14 ) { new( ptr() ) T14( std::move(t14) ); }
     variant( T15 && t15 ) : type_index( 15 ) { new( ptr() ) T15( std::move(t15) ); }
-    
+
 #endif
 
     variant(variant const & rhs)
@@ -1097,7 +1097,7 @@ public:
     variant & operator=( T13 &&      t13 ) { return move_assign_value<T13,13>( std::forward<T13>( t13 ) ); }
     variant & operator=( T14 &&      t14 ) { return move_assign_value<T14,14>( std::forward<T14>( t14 ) ); }
     variant & operator=( T15 &&      t15 ) { return move_assign_value<T15,15>( std::forward<T15>( t15 ) ); }
-    
+
 #else
     variant & operator=( T0 const & t0 ) { return copy_assign_value<T0,0>( t0 ); }
     variant & operator=( T1 const & t1 ) { return copy_assign_value<T1,1>( t1 ); }
@@ -1115,7 +1115,7 @@ public:
     variant & operator=( T13 const & t13 ) { return copy_assign_value<T13,13>( t13 ); }
     variant & operator=( T14 const & t14 ) { return copy_assign_value<T14,14>( t14 ); }
     variant & operator=( T15 const & t15 ) { return copy_assign_value<T15,15>( t15 ); }
-    
+
 #endif
 
     std::size_t index() const
@@ -1380,7 +1380,7 @@ private:
             case 13: swap( this->get<13>(), rhs.get<13>() ); break;
             case 14: swap( this->get<14>(), rhs.get<14>() ); break;
             case 15: swap( this->get<15>(), rhs.get<15>() ); break;
-            
+
         }
     }
 
@@ -1526,7 +1526,7 @@ inline Variant visit( Visitor const & vis, Variant const & v )
         case 13: return vis( get<13>( v ) );
         case 14: return vis( get<14>( v ) );
         case 15: return vis( get<15>( v ) );
-        
+
         default: return Variant();
     }
 }
@@ -1556,7 +1556,7 @@ struct Comparator
             case 13: return get<13>( v ) == get<13>( w );
             case 14: return get<14>( v ) == get<14>( w );
             case 15: return get<15>( v ) == get<15>( w );
-            
+
             default: return false;
         }
     }
@@ -1581,7 +1581,7 @@ struct Comparator
             case 13: return get<13>( v ) < get<13>( w );
             case 14: return get<14>( v ) < get<14>( w );
             case 15: return get<15>( v ) < get<15>( w );
-            
+
             default: return false;
         }
     }
